Use assisted inject

+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/michaelwiryadinatahalim/chatapp/dao/MessageDao.kt
@@ -9,5 +9,5 @@ import id.ac.ui.cs.mobileprogramming.michaelwiryadinatahalim.chatapp.model.Messa
 interface MessageDao {
 
     @Query("SELECT * FROM message_table WHERE room_chat_uid == :roomUid ORDER BY timestamp DESC")
-    fun getAllMessagesByRoomUid(roomUid: Int) : PagingSource<Int, Message>
+    fun getAllMessagesByRoomUid(roomUid: Long) : PagingSource<Int, Message>
 }
