+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/michaelwiryadinatahalim/chatapp/repositories/db/IMessageRepository.kt
@@ -5,6 +5,6 @@ import id.ac.ui.cs.mobileprogramming.michaelwiryadinatahalim.chatapp.model.Messa
 
 interface IMessageRepository {
 
-    fun getAllMessagesByRoomUid(roomUid: Int) : PagingSource<Int, Message>
+    fun getAllMessagesByRoomUid(roomUid: Long) : PagingSource<Int, Message>
 
 }
